Improved the output, now also showing the pointgroups
We show for monolayer, bilayer, and even/odd number of layers.
(Still hardcoded)

Also, might be good to add an explanation for the monolayer that it's for the symmetry, and for the bilayer that it's for the symmetry of the layer couplings.

Also added instrctions of the Fan Diagram, and upped version number to 20.11