Added the possibility to filter modes depending on the direction along which they are oscillating
This is implemented by passing along three arrays of booleans, determined in the backend,
that inform if the mode has a non-zero x, y and/or z component.
The frontend can then update the plot without reloading the data from the backend.
Also, the zoom on the y axis is fixed to the max of the y values (independent of the filter), which eases
understanding (note that zooming is still possible independently).

Also, extended the diagonalization method to use a banded algorithm.
However, it turns out this is not the problem (diagonalization takes fractions of ms, while the whole
REST API function takes 1-2 seconds for ~20 max layers).
To be further investigated (probably it's the construction of the matrix that is slow).