Added dropdowns for units
Now units are properly computed, rather than using arbitrary units
Two dropdowns allow the selection.
For the force constants, these are passed to the backend.
For the energy, everything is managed in the frontend for efficiency, assuming the backend always
returns meV (conversion is then done on the frontend).

Also improving some of the errors in case of invalid input - now the red message shows something a bit
more understandable if the user puts invalid numbers in the fields.