Adding N/m^3 as (default) force-constant units
These are better as they are used by experimentalists and are
intrisic (they do not depend on the chosen size of the unit cell).
So I'm putting those as default.
I also checked numerical values from
L. Liang et al., ACS Nano 11, 11777 (2017)
for graphite, MoS2 and h-BN; I used the values from Table I of that
paper and compared the LB and shear frequencies of the bulk with those
obtained by the tool at num_layers = 25 and they are in very good agreement,
so the unit conversion should be correct.