Just use names like `C11` for coefficients of the force-constant matrices
Originally, the idea was to call them C111 where the first of the
three indices was a layer index, for even and odd, and the last two
were Cartesian indices.
In the current implementation, we instead prepend odd- and even- to
the coefficient, so this index 1 is not needed. Simplifying the code.