Fix backscattering-active Raman modes
Some irreps were not listed (for the x and y directions),
in particular those behaving like x^2+y^2 and x^2-y^2.

This change does not have a real impact for the code, since the
active irreps for the four modes for which it matters to distinguish
x from z were already correct.

Anyway, with this commit the tables are now correct.